\name{DEU_PISA2012}
\alias{DEU_PISA2012}
\docType{data}
\title{Data from PISA 2012 - German Sample 
}
\description{
Selectetd data for 5001 'subjects' who participated in the PISA 2012 survey.
}
\usage{
  data(DEU_PISA2012)
}

\format{
  A list containing ... .
}

\details{
The data is based on freely down loadable data on the official OECD page - see source. The general structure of the data in list format, is described in an PDF document available in the User guides, package vignettes and other documentation section. 
}

\references{
Database - PISA 2003, \emph{Downloadable Data}, \verb{https://www.oecd.org/pisa/data/pisa2012database-downloadabledata.htm}
}

\examples{
##############################################################
data(DEU_PISA2012)
str(DEU_PISA2012)
}

\keyword{datasets}
