% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{induced}
\alias{induced}
\title{Transform data frame to induced graph.}
\usage{
induced(data)
}
\arguments{
\item{data}{data frame of pairwise comparison data.}
}
\value{
The induced graph.
}
\description{
Transform data frame to induced graph.
}
