% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inf_to_zero}
\alias{inf_to_zero}
\title{Replace Infinite with 0.}
\usage{
inf_to_zero(x)
}
\arguments{
\item{x}{Vector of inputs.}
}
\value{
Vector with `0`s replacing `Inf`s.
}
\description{
Replace Infinite with 0.
}
