% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{influence}
\alias{influence}
\title{Calculate influence of an observation.}
\usage{
influence(obj, source, target, sigma = 1, binary = FALSE, normalized = TRUE)
}
\arguments{
\item{obj}{Fitted \code{pairwise} object.}

\item{source, target}{The source and target nodes.}

\item{sigma}{The idiosyncratic standard deviation.}

\item{binary}{If \code{TRUE}, the question is binary. Continuous otherwise.}

\item{normalized}{If \code{TRUE} returns the old Kirchhoff index divided by the
old Kirchhoff index. If \code{FALSE}, returns the new Kirchhoff index.}
}
\value{
New normalized or non-normalized Kirchhoff index.
}
\description{
Estimates the Kirchhoff coefficient when new observations are added to
the data matrix.
}
