% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_induced_connected}
\alias{is_induced_connected}
\title{Check if the induced graph is  connected.}
\usage{
is_induced_connected(data, mode = "strong")
}
\arguments{
\item{data}{A data frame of pairwise comparison data.}

\item{mode}{Passed to \code{igraph::is.connected}. Defaults to strong.}
}
\value{
\code{TRUE} if the induced graph is strongly connected, \code{FALSE} if not.
}
\description{
Check if the induced graph is  connected.
}
