% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.pers.R
\name{lrtest.pers}
\alias{lrtest.pers}
\title{Likelihood Ratio Test for Object of class "pers"}
\usage{
lrtest.pers(object, ...)
}
\arguments{
\item{object}{an object of class\code{"pers"} - see function \code{\link{pers}}.}

\item{...}{not used jet.}
}
\description{
Function to perform a likelihood ratio test for the estimated model 'against' the saturated model for object of class\code{"pers"}.
}
