% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.pairwise}
\alias{plot.pairwise}
\title{Plot connection strength graph of a fitted pairwise model.}
\usage{
\method{plot}{pairwise}(
  x,
  reference = NULL,
  induced = TRUE,
  col = "black",
  labels = seq(obj$p),
  ...
)
}
\arguments{
\item{x}{Fitted pairwise model.}

\item{reference}{Reference node.}

\item{induced}{If \code{TRUE}, plots induced connections strengths in addition
to the main connection strengths.}

\item{col}{Passed to \code{col}.}

\item{labels}{Labels of nodes. Only used if \code{reference} is \code{NULL}.}

\item{...}{Ignored.}
}
\description{
Plot connection strength graph of a fitted pairwise model.
}
