% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{r_matrix}
\alias{r_matrix}
\title{Matrix of resistance distances.}
\usage{
r_matrix(mat)
}
\arguments{
\item{mat}{Moore-Penrose inverse of the Laplacian matrix.}
}
\value{
The resistance distance matrix R.
}
\description{
Matrix of resistance distances.
}
