% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.pers.R
\name{residuals.pers}
\alias{residuals.pers}
\title{S3 residuals for Object of class "pers"}
\usage{
\method{residuals}{pers}(object, res = "sr", na_treat = 0, ...)
}
\arguments{
\item{object}{object of class\code{"pers"}}

\item{res}{a character string defining which type of (rasch–) residual to return. This must be (exactly) one of the strings "exp" for expected scores "sr" for score residuals (default), "stdr" for standardised residuals, "srsq" for score residuals squared, or "stdrsq" for standardised residuals squared. The default is set to res="sr".}

\item{na_treat}{value to be assigned to residual cells which have missing data in the original response matrix. Default is set to na_treat=0 to set the residuals to 0, which implys that they are imputed as 'fitting data', i.e., zero residuals. This can attenuate contrasts (see. http://www.rasch.org/rmt/rmt142m.htm). An option is to set it to na_treat=NA.}

\item{...}{not used jet.}
}
\description{
S3 residuals method to extract the (Rasch) residuals for object of class\code{"pers"}
}
\keyword{methods}
