% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate}
\alias{simulate}
\title{Simulate incidence matrix.}
\usage{
simulate(clusters, vertices, conn)
}
\arguments{
\item{clusters}{Number of clusters. Defaults to `3`.}

\item{vertices}{Number of vertices per cluster. Defaults to `8`.}

\item{conn}{Symmetric matrix with number of vertices between each cluster.}
}
\value{
Simulated data.
}
\description{
Simulate incidence matrix.
}
