% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulate incidence matrix.}
\usage{
simulation(clusters, vertices, conn)
}
\arguments{
\item{clusters}{Number of clusters. Defaults to \code{3}.}

\item{vertices}{Number of vertices per cluster. Defaults to \code{8}.}

\item{conn}{Symmetric matrix with number of vertices between each cluster.}
}
\value{
Simulated data.
}
\description{
Simulate incidence matrix.
}
