% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sing_to_cov}
\alias{sing_to_cov}
\title{Calculate asymptotic covariance matrix from singular covariane}
\usage{
sing_to_cov(mat, i, keep_i = FALSE)
}
\arguments{
\item{mat}{Singular covariance matrix.}

\item{i}{Index of the baseline variable.}

\item{keep_i}{If `TRUE`, returns a singular covariance matrix with `0` in
the appropriate position.}
}
\value{
Covariance matrix with said baseline variable.
}
\description{
Calculate asymptotic covariance matrix from singular covariane
}
