% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.grm.R
\name{summary.grm}
\alias{summary.grm}
\title{S3 Summary for graphical Model Check}
\usage{
\method{summary}{grm}(object, ci = 2, ...)
}
\arguments{
\item{object}{object of class\code{"grm"}}

\item{ci}{numeric with default \code{ci=2} to return cinfidence intervalls for point estimator.}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"grm"}
}
