% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pair.S.R
\name{summary.pair.S}
\alias{summary.pair.S}
\title{S3 Summary for S-statistic Test (Wald Test)}
\usage{
\method{summary}{pair.S}(object, thres = TRUE, ...)
}
\arguments{
\item{object}{object of class\code{"pair.S"}}

\item{thres}{logical whether to output results based on the thresholds}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"pair.S"}
}
