% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pairS.R
\name{summary.pairS}
\alias{summary.pairS}
\title{S3 Summary for S-Statistic Test (Wald Test)}
\usage{
\method{summary}{pairS}(object, thres = TRUE, ...)
}
\arguments{
\item{object}{object of class\code{"pairS"}}

\item{thres}{logical whether to output results based on the thresholds}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"pairS"}
}
\keyword{methods}
