% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pairSE.R
\name{summary.pairSE}
\alias{summary.pairSE}
\title{S3 Summary for item parameter with standard errors}
\usage{
\method{summary}{pairSE}(object, sortdif = FALSE, ...)
}
\arguments{
\item{object}{object of class\code{"pairSE"}}

\item{sortdif}{logical with default \code{sortdif=FALSE} wether to order items by difficulty.}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"pairSE"}
}

