% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pairwise.person.fit.R
\name{summary.pairwise.person.fit}
\alias{summary.pairwise.person.fit}
\title{S3 Summary for Person-Fit-Statistics}
\usage{
\method{summary}{pairwise.person.fit}(object, sort = FALSE,
  by = "INFIT.ZSTD", decreasing = FALSE, ...)
}
\arguments{
\item{object}{object of class\code{"pairwise.person.fit", "data.frame" }}

\item{sort}{logical with default \code{sort=FALSE} - if set to \code{sort=TRUE} persons are ordered by absolute FIT.}

\item{by}{character passing the type of Fit-Statistic to sort by - ignored when \code{sort=FALSE}. valid options are: \code{"INFIT.ZSTD"} (default), \code{"OUTFIT.MSQ"}, \code{"OUTFIT.ZSTD"} and \code{"INFIT.MSQ"}.}

\item{decreasing}{see \code{\link{order}}}

\item{...}{other parameters passed trough - see \code{\link{order}}}
}
\description{
S3 summary method for object of class\code{c("pairwise.item.fit", "data.frame" )}
}

