% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pairwise.SepRel.R
\name{summary.pairwiseSepRel}
\alias{summary.pairwiseSepRel}
\title{S3 Summary for Person Separation Reliability}
\usage{
\method{summary}{pairwiseSepRel}(object, ...)
}
\arguments{
\item{object}{object of class\code{"pairwiseSepRel"}}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"pairwiseSepRel"}
}
\keyword{methods}
