% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.q3.R
\name{summary.q3}
\alias{summary.q3}
\title{S3 Summary for Q3 Fit Statistic}
\usage{
\method{summary}{q3}(object, maxrc = 3, ...)
}
\arguments{
\item{object}{object of class\code{"q3"}}

\item{maxrc}{numerical with default \code{maxrc=3} to specify the output of the maximum number of highest residual correlations in terms of absolute value.}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"q3"}
}
\keyword{methods}
