% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rfa.R
\name{summary.rfa}
\alias{summary.rfa}
\title{S3 Summary for rasch factor analysis}
\usage{
\method{summary}{rfa}(object, sortdif = FALSE, ...)
}
\arguments{
\item{object}{object of class\code{"pair"}}

\item{sortdif}{logical with default \code{sortdif=FALSE} wether to order items / persons by difficulty / ability.}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"pair"}
}

