% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{varf}
\alias{varf}
\title{Variance function of binary regression model}
\usage{
varf(x)
}
\arguments{
\item{x}{Vector of inputs.}
}
\value{
Variance at that vector.
}
\description{
Variance function of binary regression model
}
