% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{w_new}
\alias{w_new}
\title{Construct w.}
\usage{
w_new(d_new, beta0, sigma0)
}
\arguments{
\item{d_new}{New data.}

\item{beta0}{Vector of regression coefficients.}

\item{sigma0}{The standard deviation.}
}
\description{
Construct w.
}
