\name{cogBOOKLET}
\alias{cogBOOKLET}
\docType{data}
\title{Booklet allocation table for Math PISA (2003) data
}
\description{
a \code{data.frame} containing a booklet allocation table for the cognitive Data \code{\link{cog}} in this package, which holds 31 dichotomous items from the math task from the german sample of the PISA 2003 survey.
}
\usage{
  data(cogBOOKLET)
}

\format{
  A \code{data.frame} containing 31 rows.
}

\details{
For further Information on variables and their meaning see the codebook PDF file 
available at \verb{https://www.oecd.org/pisa/pisaproducts/PISA12_cogn_codebook.pdf}
}

\references{
Database - PISA 2003, \emph{Downloadable Data}, \verb{https://www.oecd.org/pisa/data/pisa2012database-downloadabledata.htm}
}
\examples{
data(cogBOOKLET)
cogBOOKLET
}
\keyword{datasets}
