% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\title{Fit a pairwise comparison model}
\usage{
pairwise(data, maxit = 25, epsilon = 1e-08, start = NULL)
}
\arguments{
\item{data}{Data frame to estimate.}

\item{maxit}{Maximum number of iterations. Defaults to \code{25}, which coincides
with that of \code{glm}.}

\item{epsilon}{Congvergence tolerance. Iterations converge when the relative
difference is less than epsilon.}

\item{start}{Optional starting value.}
}
\description{
Fit a pairwise comparison model
}
