% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.pairwise}
\alias{plot.pairwise}
\title{Plot connection strength graph of a fitted pairwise model.}
\usage{
\method{plot}{pairwise}(obj, reference = NULL, induced = TRUE, col = "black", ...)
}
\arguments{
\item{obj}{Fitted pairwise model.}

\item{induced}{If `TRUE`, plots induced connections strengths in addition
to the main connection strengths.}

\item{col}{Passed to `col`.}

\item{...}{Passed to `plot`.}
}
\description{
Plot connection strength graph of a fitted pairwise model.
}
