% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pers.R
\name{plot.pers}
\alias{plot.pers}
\title{S3 Plotting Person - Item Map}
\usage{
\method{plot}{pers}(x, ra = NULL, sortdif = FALSE, main = NULL,
  ylab = "Logits", itemNames = TRUE, fillCol = "grey60",
  lineCol = "grey40", cex = 0.7, pos = 4, breaks = "Sturges", pch = 1,
  ...)
}
\arguments{
\item{x}{object of class\code{"pers"}}

\item{ra}{an integer, defining the (logit) range for y-axis}

\item{sortdif}{logical wether to order items by difficulty}

\item{main}{see \code{\link{plot}}}

\item{ylab}{see \code{\link{plot}}}

\item{itemNames}{logical wether to use itemnames in the resulting plot}

\item{fillCol}{color for bar filling of the ability histogram}

\item{lineCol}{color for bar lines of the ability histogram}

\item{cex}{see \code{\link{text}}}

\item{pos}{see \code{\link{text}}}

\item{breaks}{see \code{\link{hist}}}

\item{pch}{see \code{\link{points}}}

\item{...}{other parameters passed to \code{\link{hist}} and \code{\link{text}}.}
}
\description{
S3 plotting method for object of class\code{"pers"}
}

