\name{summary.pers}
\alias{summary.pers}
\title{S3 Summary for Thetas}
\usage{
  \method{summary}{pers} (object, sortwle = FALSE, ...)
}
\arguments{
  \item{object}{object of class\code{"pers"}}

  \item{sortwle}{logical wether to order persons by
  ability}

  \item{...}{other parameters passed trough}
}
\description{
  S3 summary Method for object of class\code{"pers"}
}

